requirements.txt
asgiref==3.8.1
attrs==25.1.0
colorama==0.4.6
Django==4.2.19
django-bootstrap5==22.2
django-debug-toolbar==5.0.1
exceptiongroup==1.2.2
Faker==12.0.1
flake8==5.0.4
flake8-docstrings==1.7.0
iniconfig==2.0.0
mccabe==0.7.0
mixer==7.2.2
packaging==24.2
pep8-naming==0.13.3
pillow==11.1.0
pluggy==1.5.0
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
pytest==7.1.3
pytest-django==4.5.2
pytest-pythonpath==0.7.3
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
snowballstemmer==2.2.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
